# api_Getway/docs/swagger.yaml
basePath: /
definitions:
  authors.Author:
    properties:
      biography:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  authors.CreateAuthorRequest:
    properties:
      biography:
        type: string
      name:
        type: string
    type: object
  authors.CreateAuthorResponse:
    properties:
      author:
        $ref: '#/definitions/authors.Author'
    type: object
  authors.GetAuthorByIdResponse:
    properties:
      author:
        $ref: '#/definitions/authors.Author'
    type: object
  authors.GetAuthorsResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/authors.Author'
        type: array
    type: object
  authors.UpdateAuthorRequest:
    properties:
      biography:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  authors.UpdateAuthorResponse:
    properties:
      author:
        $ref: '#/definitions/authors.Author'
    type: object
  books.Book:
    properties:
      author_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      genre_id:
        type: string
      id:
        type: string
      summary:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  books.CreateBookRequest:
    properties:
      author_id:
        type: string
      genre_id:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
  books.CreateBookResponse:
    properties:
      book:
        $ref: '#/definitions/books.Book'
    type: object
  books.GetBookByIdResponse:
    properties:
      book:
        $ref: '#/definitions/books.Book'
    type: object
  books.GetBooksResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/books.Book'
        type: array
    type: object
  books.UpdateBookRequest:
    properties:
      author_id:
        type: string
      genre_id:
        type: string
      id:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
  books.UpdateBookResponse:
    properties:
      book:
        $ref: '#/definitions/books.Book'
    type: object
...
info:
  contact: {}
  description: API for managing Library Management System resources
  title: Library Management System API
  version: "1.0"
paths:
  /api/authors/:
    get:
      description: Retrieve a list of all authors
      parameters:
      - description: Author Name
        in: query
        name: name
        type: string
      - description: Author Biography
        in: query
        name: biography
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authors.GetAuthorsResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: Add a new author to the library system
      parameters:
      - description: Author Data
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/authors.CreateAuthorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/authors.CreateAuthorResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new author
      tags:
      - Authors
  /api/authors/{id}:
    delete:
      description: Remove an author from the library system by their ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Author deleted successfully
          schema:
            type: string
        "400":
          description: Invalid Author ID
          schema:
            type: string
        "404":
          description: Author Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete author by ID
      tags:
      - Authors
    get:
      description: Retrieve details of an author by their ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authors.GetAuthorByIdResponse'
        "400":
          description: Invalid Author ID
          schema:
            type: string
        "404":
          description: Author Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get author by ID
      tags:
      - Authors
    put:
      consumes:
      - application/json
      description: Update details of an existing author by their ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      - description: Author Data
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/authors.UpdateAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authors.UpdateAuthorResponse'
        "400":
          description: Invalid Author ID
          schema:
            type: string
        "404":
          description: Author Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update author by ID
      tags:
      - Authors
  /api/books/:
    get:
      description: Retrieve a list of all books
      parameters:
      - description: Book title
        in: query
        name: title
        type: string
      - description: Book author_id
        in: query
        name: author_id
        type: string
      - description: Book genre_id
        in: query
        name: genre_id
        type: string
      - description: Book summary
        in: query
        name: summary
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/books.GetBooksResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Add a new book to the library system
      parameters:
      - description: Book Data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/books.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/books.CreateBookResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new book
      tags:
      - Books
  /api/books/{id}:
    delete:
      description: Remove a book from the library system by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book deleted successfully
          schema:
            type: string
        "400":
          description: Invalid Book ID
          schema:
            type: string
        "404":
          description: Book Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete book by ID
      tags:
      - Books
    get:
      description: Retrieve details of a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/books.GetBookByIdResponse'
        "400":
          description: Invalid Book ID
          schema:
            type: string
        "404":
          description: Book Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get book by ID
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Update details of an existing book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Book Data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/books.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/books.UpdateBookResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "404":
          description: Book Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update book by ID
      tags:
      - Books
  /api/books/authors/{author_id}/books:
    get:
      description: Retrieve a list of books by a specific author
      parameters:
      - description: Author ID
        in: path
        name: author_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/books.GetBooksResponse'
        "400":
          description: Invalid Author ID
          schema:
            type: string
        "404":
          description: Books Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get books by author ID
      tags:
      - Books
  /api/books/genres/{genre_id}/books:
    get:
      description: Retrieve a list of books by a specific genre
      parameters:
      - description: Genre ID
        in: path
        name: genre_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/books.GetBooksResponse'
        "400":
          description: Invalid Genre ID
          schema:
            type: string
        "404":
          description: Books Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get books by genre ID
      tags:
      - Books
  /api/books/overdue:
    get:
      description: Retrieve a list of overdue books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/books.GetBooksResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get overdue books
      tags:
      - Books
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"